check authenticated user in create vote, delete vote and user disconnect

+++ src/main/java/com/influans/sp/service/VoteService.java
@@ -3,17 +3,13 @@ package com.influans.sp.service;
 import com.influans.sp.dto.DefaultResponse;
 import com.influans.sp.dto.VoteCreationDto;
 import com.influans.sp.dto.VoteDto;
-import com.influans.sp.entity.UserEntity;
 import com.influans.sp.entity.VoteEntity;
 import com.influans.sp.enums.WsTypes;
 import com.influans.sp.exception.CustomErrorCode;
 import com.influans.sp.exception.CustomException;
-import com.influans.sp.repository.SessionRepository;
 import com.influans.sp.repository.StoryRepository;
-import com.influans.sp.repository.UserRepository;
 import com.influans.sp.repository.VoteRepository;
 import com.influans.sp.security.Principal;
-import com.influans.sp.security.SecurityContext;
 import com.influans.sp.utils.StringUtils;
 import com.influans.sp.websocket.WebSocketSender;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,7 +18,6 @@ import org.springframework.stereotype.Service;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 
 /**
  * @author hazem
@@ -35,13 +30,9 @@ public class VoteService {
     @Autowired
     private StoryRepository storyRepository;
     @Autowired
-    private SessionRepository sessionRepository;
-    @Autowired
-    private UserRepository userRepository;
-    @Autowired
     private WebSocketSender webSocketSender;
     @Autowired
-    private SecurityContext securityContext;
+    private AuthenticationService authenticationService;
 
     /**
      * @param storyId storyId
@@ -73,13 +64,16 @@ public class VoteService {
     /**
      * @param voteId voteId
      * @return empty response
-     * @should throw an exception if user is not connected to the related session
+     * @should check authenticated user
      * @should throw an exception if voteId is null
      * @should throw an exception if vote does not exist with given id
+     * @should throw an exception if user is not the vote owner
      * @should delete vote with the given id
      * @should send a websocket notification
      */
     public DefaultResponse delete(String voteId) {
+        final Principal user = authenticationService.checkAuthenticatedUser();
+
         if (StringUtils.isEmpty(voteId)) {
             throw new CustomException(CustomErrorCode.BAD_ARGS, "voteId should not be null or empty");
         }
@@ -90,6 +84,11 @@ public class VoteService {
             throw new CustomException(CustomErrorCode.OBJECT_NOT_FOUND, "no vote found with given Id " + voteId);
         }
 
+        if (!voteEntity.getUsername().equals(user.getUsername()) || !voteEntity.getSessionId().equals(user.getSessionId())) {
+            throw new CustomException(CustomErrorCode.PERMISSION_DENIED, "username %s is not permitted to delete vote %s in session %s",
+                    user.getUsername(), voteId, user.getSessionId());
+        }
+
         voteRepository.delete(voteId);
         webSocketSender.sendNotification(voteEntity.getSessionId(), WsTypes.VOTE_REMOVED, voteId);
         return DefaultResponse.ok();
@@ -98,58 +97,27 @@ public class VoteService {
     /**
      * @param voteCreationDto voteDto
      * @return voteDto with new id
+     * @should check authenticated user
      * @should throw an exception if storyId is null or empty
-     * @should throw an exception if sessionId is null or empty
-     * @should throw an exception if username is null or empty
      * @should throw an exception if value is null or empty
      * @should throw an exception if story does not exist with given Id
-     * @should throw an exception if no user has been connected to the related session with the given username
-     * @should throw an exception if user has been disconnected from the related session
-     * @should throw an exception if session does not exist with given sessionId
      * @should Update existing vote if the user has already voted on the given story
      * @should create a vote for the given user on the selected story
      * @should send a websocket notification
      */
     public VoteCreationDto saveVote(VoteCreationDto voteCreationDto) {
-        final Optional<Principal> optional = securityContext.getAuthenticationContext();
-
-        if (!optional.isPresent()) {
-            throw new CustomException(CustomErrorCode.UNAUTHORIZED, "user not authenticated");
-        }
-
-        final Principal user = optional.get();
-
-        if (StringUtils.isEmpty(user.getSessionId())) {
-            throw new CustomException(CustomErrorCode.UNAUTHORIZED, "sessionId should not be null or empty");
-        }
+        final Principal user = authenticationService.checkAuthenticatedUser();
 
         if (StringUtils.isEmpty(voteCreationDto.getStoryId())) {
             throw new CustomException(CustomErrorCode.BAD_ARGS, "storyId should not be null or empty");
-        }
-
-        if (StringUtils.isEmpty(user.getUsername())) {
-            throw new CustomException(CustomErrorCode.UNAUTHORIZED, "username should not be null or empty");
         }
 
         if (StringUtils.isEmpty(voteCreationDto.getValue())) {
             throw new CustomException(CustomErrorCode.BAD_ARGS, "value should not be null or empty");
         }
 
-        if (!sessionRepository.exists(user.getSessionId())) {
-            throw new CustomException(CustomErrorCode.UNAUTHORIZED, "session not found with id = " + user.getSessionId());
-        }
-
         if (!storyRepository.exists(voteCreationDto.getStoryId())) {
             throw new CustomException(CustomErrorCode.OBJECT_NOT_FOUND, "story not found with id = " + voteCreationDto.getStoryId());
-        }
-
-        final UserEntity userEntity = userRepository.findUser(user.getSessionId(), user.getUsername());
-        if (userEntity == null) {
-            throw new CustomException(CustomErrorCode.UNAUTHORIZED, "user not found with username %s in session %s ",
-                    user.getUsername(), user.getSessionId());
-        } else if (!userEntity.isConnected()) {
-            throw new CustomException(CustomErrorCode.UNAUTHORIZED, "user %s has been disconnected  from session %s ",
-                    user.getUsername(), user.getSessionId());
         }
 
         VoteEntity voteEntity = voteRepository.getVoteByUserOnStory(user.getUsername(), voteCreationDto.getStoryId());
